+++ qcal/__init__.py
@@ -1,8 +1,9 @@
 """Quantum Calibration (qcal) package."""
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 from qcal.circuit import *
+from qcal.compilation.compiler import Compiler
 from qcal.config import Config
 from qcal.gate.gate import Gate
 from qcal.gate.single_qubit import *