+++ qcal/compilation/__init__.py
@@ -1,0 +1,3 @@
+from qcal.compilation.compiler import Compiler
+
+__all__ = ('Compiler',)